+++ src/app/view/mission/new/NewMissionDatasourceDetectionsComponent.js
@@ -16,6 +16,27 @@ class NewMissionDatasourceDetectionsComponent extends Component {
 			selectedType: "",
 			allowedEditors: [ "importer" ]
 		};
+		
+		this.typeNames = {
+			[P4C.Detection.OBJECT_BENCH]: I18n.t("Bench"),
+			[P4C.Detection.SIGN_STOP]: I18n.t("Stop sign"),
+			[P4C.Detection.MARK_CROSSING]: I18n.t("Pedestrian crossing"),
+			[P4C.Detection.OBJECT_BICYCLE_PARKING]: I18n.t("Bicycle parking"),
+			[P4C.Detection.OBJECT_CCTV]: I18n.t("CCTV camera"),
+			[P4C.Detection.OBJECT_HYDRANT]: I18n.t("Fire hydrant"),
+			[P4C.Detection.OBJECT_POSTBOX]: I18n.t("Post box"),
+			[P4C.Detection.OBJECT_MANHOLE]: I18n.t("Manhole"),
+			[P4C.Detection.OBJECT_PARKING_METER]: I18n.t("Parking meter"),
+			[P4C.Detection.OBJECT_PHONE]: I18n.t("Phone (public/emergency)"),
+			[P4C.Detection.SIGN_ADVERT]: I18n.t("Advert sign"),
+			[P4C.Detection.SIGN_INFO]: I18n.t("Information sign"),
+			[P4C.Detection.SIGN_STORE]: I18n.t("Store sign"),
+			[P4C.Detection.OBJECT_STREET_LIGHT]: I18n.t("Street light"),
+			[P4C.Detection.OBJECT_POLE]: I18n.t("Pole"),
+			[P4C.Detection.SIGN_RESERVED_PARKING]: I18n.t("Reserved parking"),
+			[P4C.Detection.SIGN_ANIMAL_CROSSING]: I18n.t("Animal crossing sign"),
+			[P4C.Detection.SIGN_RAILWAY_CROSSING]: I18n.t("Railway crossing sign")
+		};
 	}
 	
 	/**
@@ -65,7 +86,7 @@ class NewMissionDatasourceDetectionsComponent extends Component {
 			>
 				<option value="" />
 				{Object.entries(P4C.Detection.TYPE_DETAILS).map((e, i) => {
-					return <option value={e[0]} key={i}>{e[1].name}</option>;
+					return <option value={e[0]} key={i}>{this.typeNames[e[0]] || e[1].name}</option>;
 				})}
 			</Select>
 		</FormControl>;